+++ tests/testthat/test-calc_vol.R
@@ -13,7 +13,7 @@ test_that("returns correct number of columns depending on return arguments", {
 })
 
 test_that("calc_vol() works with multiple inputs", {
-  paths <- c("tests/testthat/data/map00361/C00011.mol", "tests/testthat/data/map00361/C00042.mol")
+  paths <- c("data/map00361/C00011.mol", "data/map00361/C00042.mol")
   expect_s3_class(calc_vol(paths), "data.frame")
 })